Add Auth to execution spec
