+++ gen/pb-go/flyteidl/admin/common.pb.validate.go
@@ -1597,3 +1597,77 @@ var _ interface {
 	Cause() error
 	ErrorName() string
 } = AnnotationsValidationError{}
+
+// Validate checks the field values on Auth with the rules defined in the proto
+// definition for this message. If any rules are violated, an error is returned.
+func (m *Auth) Validate() error {
+	if m == nil {
+		return nil
+	}
+
+	switch m.Method.(type) {
+
+	case *Auth_AssumableIamRole:
+		// no validation rules for AssumableIamRole
+
+	case *Auth_KubernetesServiceAccount:
+		// no validation rules for KubernetesServiceAccount
+
+	}
+
+	return nil
+}
+
+// AuthValidationError is the validation error returned by Auth.Validate if the
+// designated constraints aren't met.
+type AuthValidationError struct {
+	field  string
+	reason string
+	cause  error
+	key    bool
+}
+
+// Field function returns field value.
+func (e AuthValidationError) Field() string { return e.field }
+
+// Reason function returns reason value.
+func (e AuthValidationError) Reason() string { return e.reason }
+
+// Cause function returns cause value.
+func (e AuthValidationError) Cause() error { return e.cause }
+
+// Key function returns key value.
+func (e AuthValidationError) Key() bool { return e.key }
+
+// ErrorName returns error name.
+func (e AuthValidationError) ErrorName() string { return "AuthValidationError" }
+
+// Error satisfies the builtin error interface
+func (e AuthValidationError) Error() string {
+	cause := ""
+	if e.cause != nil {
+		cause = fmt.Sprintf(" | caused by: %v", e.cause)
+	}
+
+	key := ""
+	if e.key {
+		key = "key for "
+	}
+
+	return fmt.Sprintf(
+		"invalid %sAuth.%s: %s%s",
+		key,
+		e.field,
+		e.reason,
+		cause)
+}
+
+var _ error = AuthValidationError{}
+
+var _ interface {
+	Field() string
+	Reason() string
+	Key() bool
+	Cause() error
+	ErrorName() string
+} = AuthValidationError{}
