+++ gen/pb-go/flyteidl/admin/execution.pb.validate.go
@@ -1199,6 +1199,16 @@ func (m *ExecutionSpec) Validate() error {
 		}
 	}
 
+	if v, ok := interface{}(m.GetAuth()).(interface{ Validate() error }); ok {
+		if err := v.Validate(); err != nil {
+			return ExecutionSpecValidationError{
+				field:  "Auth",
+				reason: "embedded message failed validation",
+				cause:  err,
+			}
+		}
+	}
+
 	switch m.NotificationOverrides.(type) {
 
 	case *ExecutionSpec_Notifications:
