+++ gen/pb-go/flyteidl/service/flyteadmin/model_admin_execution_spec.go
@@ -22,4 +22,6 @@ type AdminExecutionSpec struct {
 	Labels *AdminLabels `json:"labels,omitempty"`
 	// Annotations to apply to the execution resource.
 	Annotations *AdminAnnotations `json:"annotations,omitempty"`
+	// Optional: auth override to apply this execution.
+	Auth *AdminAuth `json:"auth,omitempty"`
 }
