+++ gen/pb-java/flyteidl/admin/ExecutionOuterClass.java
@@ -14872,6 +14872,31 @@ public final class ExecutionOuterClass {
      */
     flyteidl.admin.Common.AnnotationsOrBuilder getAnnotationsOrBuilder();
 
+    /**
+     * <pre>
+     * Optional: auth override to apply this execution.
+     * </pre>
+     *
+     * <code>.flyteidl.admin.Auth auth = 16;</code>
+     */
+    boolean hasAuth();
+    /**
+     * <pre>
+     * Optional: auth override to apply this execution.
+     * </pre>
+     *
+     * <code>.flyteidl.admin.Auth auth = 16;</code>
+     */
+    flyteidl.admin.Common.Auth getAuth();
+    /**
+     * <pre>
+     * Optional: auth override to apply this execution.
+     * </pre>
+     *
+     * <code>.flyteidl.admin.Auth auth = 16;</code>
+     */
+    flyteidl.admin.Common.AuthOrBuilder getAuthOrBuilder();
+
     public flyteidl.admin.ExecutionOuterClass.ExecutionSpec.NotificationOverridesCase getNotificationOverridesCase();
   }
   /**
@@ -14998,6 +15023,19 @@ public final class ExecutionOuterClass {
               if (subBuilder != null) {
                 subBuilder.mergeFrom(annotations_);
                 annotations_ = subBuilder.buildPartial();
+              }
+
+              break;
+            }
+            case 130: {
+              flyteidl.admin.Common.Auth.Builder subBuilder = null;
+              if (auth_ != null) {
+                subBuilder = auth_.toBuilder();
+              }
+              auth_ = input.readMessage(flyteidl.admin.Common.Auth.parser(), extensionRegistry);
+              if (subBuilder != null) {
+                subBuilder.mergeFrom(auth_);
+                auth_ = subBuilder.buildPartial();
               }
 
               break;
@@ -15296,6 +15334,39 @@ public final class ExecutionOuterClass {
       return getAnnotations();
     }
 
+    public static final int AUTH_FIELD_NUMBER = 16;
+    private flyteidl.admin.Common.Auth auth_;
+    /**
+     * <pre>
+     * Optional: auth override to apply this execution.
+     * </pre>
+     *
+     * <code>.flyteidl.admin.Auth auth = 16;</code>
+     */
+    public boolean hasAuth() {
+      return auth_ != null;
+    }
+    /**
+     * <pre>
+     * Optional: auth override to apply this execution.
+     * </pre>
+     *
+     * <code>.flyteidl.admin.Auth auth = 16;</code>
+     */
+    public flyteidl.admin.Common.Auth getAuth() {
+      return auth_ == null ? flyteidl.admin.Common.Auth.getDefaultInstance() : auth_;
+    }
+    /**
+     * <pre>
+     * Optional: auth override to apply this execution.
+     * </pre>
+     *
+     * <code>.flyteidl.admin.Auth auth = 16;</code>
+     */
+    public flyteidl.admin.Common.AuthOrBuilder getAuthOrBuilder() {
+      return getAuth();
+    }
+
     private byte memoizedIsInitialized = -1;
     @java.lang.Override
     public final boolean isInitialized() {
@@ -15331,6 +15402,9 @@ public final class ExecutionOuterClass {
       }
       if (annotations_ != null) {
         output.writeMessage(8, getAnnotations());
+      }
+      if (auth_ != null) {
+        output.writeMessage(16, getAuth());
       }
       unknownFields.writeTo(output);
     }
@@ -15369,6 +15443,10 @@ public final class ExecutionOuterClass {
       if (annotations_ != null) {
         size += com.google.protobuf.CodedOutputStream
           .computeMessageSize(8, getAnnotations());
+      }
+      if (auth_ != null) {
+        size += com.google.protobuf.CodedOutputStream
+          .computeMessageSize(16, getAuth());
       }
       size += unknownFields.getSerializedSize();
       memoizedSize = size;
@@ -15409,6 +15487,11 @@ public final class ExecutionOuterClass {
       if (hasAnnotations()) {
         if (!getAnnotations()
             .equals(other.getAnnotations())) return false;
+      }
+      if (hasAuth() != other.hasAuth()) return false;
+      if (hasAuth()) {
+        if (!getAuth()
+            .equals(other.getAuth())) return false;
       }
       if (!getNotificationOverridesCase().equals(other.getNotificationOverridesCase())) return false;
       switch (notificationOverridesCase_) {
@@ -15453,6 +15536,10 @@ public final class ExecutionOuterClass {
       if (hasAnnotations()) {
         hash = (37 * hash) + ANNOTATIONS_FIELD_NUMBER;
         hash = (53 * hash) + getAnnotations().hashCode();
+      }
+      if (hasAuth()) {
+        hash = (37 * hash) + AUTH_FIELD_NUMBER;
+        hash = (53 * hash) + getAuth().hashCode();
       }
       switch (notificationOverridesCase_) {
         case 5:
@@ -15635,6 +15722,12 @@ public final class ExecutionOuterClass {
           annotations_ = null;
           annotationsBuilder_ = null;
         }
+        if (authBuilder_ == null) {
+          auth_ = null;
+        } else {
+          auth_ = null;
+          authBuilder_ = null;
+        }
         notificationOverridesCase_ = 0;
         notificationOverrides_ = null;
         return this;
@@ -15697,6 +15790,11 @@ public final class ExecutionOuterClass {
           result.annotations_ = annotations_;
         } else {
           result.annotations_ = annotationsBuilder_.build();
+        }
+        if (authBuilder_ == null) {
+          result.auth_ = auth_;
+        } else {
+          result.auth_ = authBuilder_.build();
         }
         result.notificationOverridesCase_ = notificationOverridesCase_;
         onBuilt();
@@ -15761,6 +15859,9 @@ public final class ExecutionOuterClass {
         }
         if (other.hasAnnotations()) {
           mergeAnnotations(other.getAnnotations());
+        }
+        if (other.hasAuth()) {
+          mergeAuth(other.getAuth());
         }
         switch (other.getNotificationOverridesCase()) {
           case NOTIFICATIONS: {
@@ -16814,6 +16915,159 @@ public final class ExecutionOuterClass {
           annotations_ = null;
         }
         return annotationsBuilder_;
+      }
+
+      private flyteidl.admin.Common.Auth auth_;
+      private com.google.protobuf.SingleFieldBuilderV3<
+          flyteidl.admin.Common.Auth, flyteidl.admin.Common.Auth.Builder, flyteidl.admin.Common.AuthOrBuilder> authBuilder_;
+      /**
+       * <pre>
+       * Optional: auth override to apply this execution.
+       * </pre>
+       *
+       * <code>.flyteidl.admin.Auth auth = 16;</code>
+       */
+      public boolean hasAuth() {
+        return authBuilder_ != null || auth_ != null;
+      }
+      /**
+       * <pre>
+       * Optional: auth override to apply this execution.
+       * </pre>
+       *
+       * <code>.flyteidl.admin.Auth auth = 16;</code>
+       */
+      public flyteidl.admin.Common.Auth getAuth() {
+        if (authBuilder_ == null) {
+          return auth_ == null ? flyteidl.admin.Common.Auth.getDefaultInstance() : auth_;
+        } else {
+          return authBuilder_.getMessage();
+        }
+      }
+      /**
+       * <pre>
+       * Optional: auth override to apply this execution.
+       * </pre>
+       *
+       * <code>.flyteidl.admin.Auth auth = 16;</code>
+       */
+      public Builder setAuth(flyteidl.admin.Common.Auth value) {
+        if (authBuilder_ == null) {
+          if (value == null) {
+            throw new NullPointerException();
+          }
+          auth_ = value;
+          onChanged();
+        } else {
+          authBuilder_.setMessage(value);
+        }
+
+        return this;
+      }
+      /**
+       * <pre>
+       * Optional: auth override to apply this execution.
+       * </pre>
+       *
+       * <code>.flyteidl.admin.Auth auth = 16;</code>
+       */
+      public Builder setAuth(
+          flyteidl.admin.Common.Auth.Builder builderForValue) {
+        if (authBuilder_ == null) {
+          auth_ = builderForValue.build();
+          onChanged();
+        } else {
+          authBuilder_.setMessage(builderForValue.build());
+        }
+
+        return this;
+      }
+      /**
+       * <pre>
+       * Optional: auth override to apply this execution.
+       * </pre>
+       *
+       * <code>.flyteidl.admin.Auth auth = 16;</code>
+       */
+      public Builder mergeAuth(flyteidl.admin.Common.Auth value) {
+        if (authBuilder_ == null) {
+          if (auth_ != null) {
+            auth_ =
+              flyteidl.admin.Common.Auth.newBuilder(auth_).mergeFrom(value).buildPartial();
+          } else {
+            auth_ = value;
+          }
+          onChanged();
+        } else {
+          authBuilder_.mergeFrom(value);
+        }
+
+        return this;
+      }
+      /**
+       * <pre>
+       * Optional: auth override to apply this execution.
+       * </pre>
+       *
+       * <code>.flyteidl.admin.Auth auth = 16;</code>
+       */
+      public Builder clearAuth() {
+        if (authBuilder_ == null) {
+          auth_ = null;
+          onChanged();
+        } else {
+          auth_ = null;
+          authBuilder_ = null;
+        }
+
+        return this;
+      }
+      /**
+       * <pre>
+       * Optional: auth override to apply this execution.
+       * </pre>
+       *
+       * <code>.flyteidl.admin.Auth auth = 16;</code>
+       */
+      public flyteidl.admin.Common.Auth.Builder getAuthBuilder() {
+        
+        onChanged();
+        return getAuthFieldBuilder().getBuilder();
+      }
+      /**
+       * <pre>
+       * Optional: auth override to apply this execution.
+       * </pre>
+       *
+       * <code>.flyteidl.admin.Auth auth = 16;</code>
+       */
+      public flyteidl.admin.Common.AuthOrBuilder getAuthOrBuilder() {
+        if (authBuilder_ != null) {
+          return authBuilder_.getMessageOrBuilder();
+        } else {
+          return auth_ == null ?
+              flyteidl.admin.Common.Auth.getDefaultInstance() : auth_;
+        }
+      }
+      /**
+       * <pre>
+       * Optional: auth override to apply this execution.
+       * </pre>
+       *
+       * <code>.flyteidl.admin.Auth auth = 16;</code>
+       */
+      private com.google.protobuf.SingleFieldBuilderV3<
+          flyteidl.admin.Common.Auth, flyteidl.admin.Common.Auth.Builder, flyteidl.admin.Common.AuthOrBuilder> 
+          getAuthFieldBuilder() {
+        if (authBuilder_ == null) {
+          authBuilder_ = new com.google.protobuf.SingleFieldBuilderV3<
+              flyteidl.admin.Common.Auth, flyteidl.admin.Common.Auth.Builder, flyteidl.admin.Common.AuthOrBuilder>(
+                  getAuth(),
+                  getParentForChildren(),
+                  isClean());
+          auth_ = null;
+        }
+        return authBuilder_;
       }
       @java.lang.Override
       public final Builder setUnknownFields(
@@ -19894,7 +20148,7 @@ public final class ExecutionOuterClass {
       "CHEDULED\020\001\022\n\n\006SYSTEM\020\002\022\014\n\010RELAUNCH\020\003\022\022\n\016" +
       "CHILD_WORKFLOW\020\004\"G\n\020NotificationList\0223\n\r" +
       "notifications\030\001 \003(\0132\034.flyteidl.admin.Not" +
-      "ification\"\357\002\n\rExecutionSpec\022.\n\013launch_pl" +
+      "ification\"\223\003\n\rExecutionSpec\022.\n\013launch_pl" +
       "an\030\001 \001(\0132\031.flyteidl.core.Identifier\022-\n\006i" +
       "nputs\030\002 \001(\0132\031.flyteidl.core.LiteralMapB\002" +
       "\030\001\0223\n\010metadata\030\003 \001(\0132!.flyteidl.admin.Ex" +
@@ -19902,18 +20156,19 @@ public final class ExecutionOuterClass {
       ".flyteidl.admin.NotificationListH\000\022\025\n\013di" +
       "sable_all\030\006 \001(\010H\000\022&\n\006labels\030\007 \001(\0132\026.flyt" +
       "eidl.admin.Labels\0220\n\013annotations\030\010 \001(\0132\033" +
-      ".flyteidl.admin.AnnotationsB\030\n\026notificat" +
-      "ion_overridesJ\004\010\004\020\005\"b\n\031ExecutionTerminat" +
-      "eRequest\0226\n\002id\030\001 \001(\0132*.flyteidl.core.Wor" +
-      "kflowExecutionIdentifier\022\r\n\005cause\030\002 \001(\t\"" +
-      "\034\n\032ExecutionTerminateResponse\"Y\n\037Workflo" +
-      "wExecutionGetDataRequest\0226\n\002id\030\001 \001(\0132*.f" +
-      "lyteidl.core.WorkflowExecutionIdentifier" +
-      "\"u\n WorkflowExecutionGetDataResponse\022(\n\007" +
-      "outputs\030\001 \001(\0132\027.flyteidl.admin.UrlBlob\022\'" +
-      "\n\006inputs\030\002 \001(\0132\027.flyteidl.admin.UrlBlobB" +
-      "3Z1github.com/lyft/flyteidl/gen/pb-go/fl" +
-      "yteidl/adminb\006proto3"
+      ".flyteidl.admin.Annotations\022\"\n\004auth\030\020 \001(" +
+      "\0132\024.flyteidl.admin.AuthB\030\n\026notification_" +
+      "overridesJ\004\010\004\020\005\"b\n\031ExecutionTerminateReq" +
+      "uest\0226\n\002id\030\001 \001(\0132*.flyteidl.core.Workflo" +
+      "wExecutionIdentifier\022\r\n\005cause\030\002 \001(\t\"\034\n\032E" +
+      "xecutionTerminateResponse\"Y\n\037WorkflowExe" +
+      "cutionGetDataRequest\0226\n\002id\030\001 \001(\0132*.flyte" +
+      "idl.core.WorkflowExecutionIdentifier\"u\n " +
+      "WorkflowExecutionGetDataResponse\022(\n\007outp" +
+      "uts\030\001 \001(\0132\027.flyteidl.admin.UrlBlob\022\'\n\006in" +
+      "puts\030\002 \001(\0132\027.flyteidl.admin.UrlBlobB3Z1g" +
+      "ithub.com/lyft/flyteidl/gen/pb-go/flytei" +
+      "dl/adminb\006proto3"
     };
     com.google.protobuf.Descriptors.FileDescriptor.InternalDescriptorAssigner assigner =
         new com.google.protobuf.Descriptors.FileDescriptor.    InternalDescriptorAssigner() {
@@ -20010,7 +20265,7 @@ public final class ExecutionOuterClass {
     internal_static_flyteidl_admin_ExecutionSpec_fieldAccessorTable = new
       com.google.protobuf.GeneratedMessageV3.FieldAccessorTable(
         internal_static_flyteidl_admin_ExecutionSpec_descriptor,
-        new java.lang.String[] { "LaunchPlan", "Inputs", "Metadata", "Notifications", "DisableAll", "Labels", "Annotations", "NotificationOverrides", });
+        new java.lang.String[] { "LaunchPlan", "Inputs", "Metadata", "Notifications", "DisableAll", "Labels", "Annotations", "Auth", "NotificationOverrides", });
     internal_static_flyteidl_admin_ExecutionTerminateRequest_descriptor =
       getDescriptor().getMessageTypes().get(13);
     internal_static_flyteidl_admin_ExecutionTerminateRequest_fieldAccessorTable = new
