+++ gen/pb-js/flyteidl.js
@@ -14464,6 +14464,154 @@ export const flyteidl = $root.flyteidl = (() => {
             return Annotations;
         })();
 
+        admin.Auth = (function() {
+
+            /**
+             * Properties of an Auth.
+             * @memberof flyteidl.admin
+             * @interface IAuth
+             * @property {string|null} [assumableIamRole] Auth assumableIamRole
+             * @property {string|null} [kubernetesServiceAccount] Auth kubernetesServiceAccount
+             */
+
+            /**
+             * Constructs a new Auth.
+             * @memberof flyteidl.admin
+             * @classdesc Represents an Auth.
+             * @implements IAuth
+             * @constructor
+             * @param {flyteidl.admin.IAuth=} [properties] Properties to set
+             */
+            function Auth(properties) {
+                if (properties)
+                    for (let keys = Object.keys(properties), i = 0; i < keys.length; ++i)
+                        if (properties[keys[i]] != null)
+                            this[keys[i]] = properties[keys[i]];
+            }
+
+            /**
+             * Auth assumableIamRole.
+             * @member {string} assumableIamRole
+             * @memberof flyteidl.admin.Auth
+             * @instance
+             */
+            Auth.prototype.assumableIamRole = "";
+
+            /**
+             * Auth kubernetesServiceAccount.
+             * @member {string} kubernetesServiceAccount
+             * @memberof flyteidl.admin.Auth
+             * @instance
+             */
+            Auth.prototype.kubernetesServiceAccount = "";
+
+            // OneOf field names bound to virtual getters and setters
+            let $oneOfFields;
+
+            /**
+             * Auth method.
+             * @member {"assumableIamRole"|"kubernetesServiceAccount"|undefined} method
+             * @memberof flyteidl.admin.Auth
+             * @instance
+             */
+            Object.defineProperty(Auth.prototype, "method", {
+                get: $util.oneOfGetter($oneOfFields = ["assumableIamRole", "kubernetesServiceAccount"]),
+                set: $util.oneOfSetter($oneOfFields)
+            });
+
+            /**
+             * Creates a new Auth instance using the specified properties.
+             * @function create
+             * @memberof flyteidl.admin.Auth
+             * @static
+             * @param {flyteidl.admin.IAuth=} [properties] Properties to set
+             * @returns {flyteidl.admin.Auth} Auth instance
+             */
+            Auth.create = function create(properties) {
+                return new Auth(properties);
+            };
+
+            /**
+             * Encodes the specified Auth message. Does not implicitly {@link flyteidl.admin.Auth.verify|verify} messages.
+             * @function encode
+             * @memberof flyteidl.admin.Auth
+             * @static
+             * @param {flyteidl.admin.IAuth} message Auth message or plain object to encode
+             * @param {$protobuf.Writer} [writer] Writer to encode to
+             * @returns {$protobuf.Writer} Writer
+             */
+            Auth.encode = function encode(message, writer) {
+                if (!writer)
+                    writer = $Writer.create();
+                if (message.assumableIamRole != null && message.hasOwnProperty("assumableIamRole"))
+                    writer.uint32(/* id 1, wireType 2 =*/10).string(message.assumableIamRole);
+                if (message.kubernetesServiceAccount != null && message.hasOwnProperty("kubernetesServiceAccount"))
+                    writer.uint32(/* id 2, wireType 2 =*/18).string(message.kubernetesServiceAccount);
+                return writer;
+            };
+
+            /**
+             * Decodes an Auth message from the specified reader or buffer.
+             * @function decode
+             * @memberof flyteidl.admin.Auth
+             * @static
+             * @param {$protobuf.Reader|Uint8Array} reader Reader or buffer to decode from
+             * @param {number} [length] Message length if known beforehand
+             * @returns {flyteidl.admin.Auth} Auth
+             * @throws {Error} If the payload is not a reader or valid buffer
+             * @throws {$protobuf.util.ProtocolError} If required fields are missing
+             */
+            Auth.decode = function decode(reader, length) {
+                if (!(reader instanceof $Reader))
+                    reader = $Reader.create(reader);
+                let end = length === undefined ? reader.len : reader.pos + length, message = new $root.flyteidl.admin.Auth();
+                while (reader.pos < end) {
+                    let tag = reader.uint32();
+                    switch (tag >>> 3) {
+                    case 1:
+                        message.assumableIamRole = reader.string();
+                        break;
+                    case 2:
+                        message.kubernetesServiceAccount = reader.string();
+                        break;
+                    default:
+                        reader.skipType(tag & 7);
+                        break;
+                    }
+                }
+                return message;
+            };
+
+            /**
+             * Verifies an Auth message.
+             * @function verify
+             * @memberof flyteidl.admin.Auth
+             * @static
+             * @param {Object.<string,*>} message Plain object to verify
+             * @returns {string|null} `null` if valid, otherwise the reason why it is not
+             */
+            Auth.verify = function verify(message) {
+                if (typeof message !== "object" || message === null)
+                    return "object expected";
+                let properties = {};
+                if (message.assumableIamRole != null && message.hasOwnProperty("assumableIamRole")) {
+                    properties.method = 1;
+                    if (!$util.isString(message.assumableIamRole))
+                        return "assumableIamRole: string expected";
+                }
+                if (message.kubernetesServiceAccount != null && message.hasOwnProperty("kubernetesServiceAccount")) {
+                    if (properties.method === 1)
+                        return "method: multiple values";
+                    properties.method = 1;
+                    if (!$util.isString(message.kubernetesServiceAccount))
+                        return "kubernetesServiceAccount: string expected";
+                }
+                return null;
+            };
+
+            return Auth;
+        })();
+
         admin.EventErrorAlreadyInTerminalState = (function() {
 
             /**
@@ -17330,6 +17478,7 @@ export const flyteidl = $root.flyteidl = (() => {
              * @property {boolean|null} [disableAll] ExecutionSpec disableAll
              * @property {flyteidl.admin.ILabels|null} [labels] ExecutionSpec labels
              * @property {flyteidl.admin.IAnnotations|null} [annotations] ExecutionSpec annotations
+             * @property {flyteidl.admin.IAuth|null} [auth] ExecutionSpec auth
              */
 
             /**
@@ -17403,6 +17552,14 @@ export const flyteidl = $root.flyteidl = (() => {
              */
             ExecutionSpec.prototype.annotations = null;
 
+            /**
+             * ExecutionSpec auth.
+             * @member {flyteidl.admin.IAuth|null|undefined} auth
+             * @memberof flyteidl.admin.ExecutionSpec
+             * @instance
+             */
+            ExecutionSpec.prototype.auth = null;
+
             // OneOf field names bound to virtual getters and setters
             let $oneOfFields;
 
@@ -17455,6 +17612,8 @@ export const flyteidl = $root.flyteidl = (() => {
                     $root.flyteidl.admin.Labels.encode(message.labels, writer.uint32(/* id 7, wireType 2 =*/58).fork()).ldelim();
                 if (message.annotations != null && message.hasOwnProperty("annotations"))
                     $root.flyteidl.admin.Annotations.encode(message.annotations, writer.uint32(/* id 8, wireType 2 =*/66).fork()).ldelim();
+                if (message.auth != null && message.hasOwnProperty("auth"))
+                    $root.flyteidl.admin.Auth.encode(message.auth, writer.uint32(/* id 16, wireType 2 =*/130).fork()).ldelim();
                 return writer;
             };
 
@@ -17496,6 +17655,9 @@ export const flyteidl = $root.flyteidl = (() => {
                         break;
                     case 8:
                         message.annotations = $root.flyteidl.admin.Annotations.decode(reader, reader.uint32());
+                        break;
+                    case 16:
+                        message.auth = $root.flyteidl.admin.Auth.decode(reader, reader.uint32());
                         break;
                     default:
                         reader.skipType(tag & 7);
@@ -17556,6 +17718,11 @@ export const flyteidl = $root.flyteidl = (() => {
                     let error = $root.flyteidl.admin.Annotations.verify(message.annotations);
                     if (error)
                         return "annotations." + error;
+                }
+                if (message.auth != null && message.hasOwnProperty("auth")) {
+                    let error = $root.flyteidl.admin.Auth.verify(message.auth);
+                    if (error)
+                        return "auth." + error;
                 }
                 return null;
             };
@@ -18551,154 +18718,6 @@ export const flyteidl = $root.flyteidl = (() => {
             };
 
             return LaunchPlanList;
-        })();
-
-        admin.Auth = (function() {
-
-            /**
-             * Properties of an Auth.
-             * @memberof flyteidl.admin
-             * @interface IAuth
-             * @property {string|null} [assumableIamRole] Auth assumableIamRole
-             * @property {string|null} [kubernetesServiceAccount] Auth kubernetesServiceAccount
-             */
-
-            /**
-             * Constructs a new Auth.
-             * @memberof flyteidl.admin
-             * @classdesc Represents an Auth.
-             * @implements IAuth
-             * @constructor
-             * @param {flyteidl.admin.IAuth=} [properties] Properties to set
-             */
-            function Auth(properties) {
-                if (properties)
-                    for (let keys = Object.keys(properties), i = 0; i < keys.length; ++i)
-                        if (properties[keys[i]] != null)
-                            this[keys[i]] = properties[keys[i]];
-            }
-
-            /**
-             * Auth assumableIamRole.
-             * @member {string} assumableIamRole
-             * @memberof flyteidl.admin.Auth
-             * @instance
-             */
-            Auth.prototype.assumableIamRole = "";
-
-            /**
-             * Auth kubernetesServiceAccount.
-             * @member {string} kubernetesServiceAccount
-             * @memberof flyteidl.admin.Auth
-             * @instance
-             */
-            Auth.prototype.kubernetesServiceAccount = "";
-
-            // OneOf field names bound to virtual getters and setters
-            let $oneOfFields;
-
-            /**
-             * Auth method.
-             * @member {"assumableIamRole"|"kubernetesServiceAccount"|undefined} method
-             * @memberof flyteidl.admin.Auth
-             * @instance
-             */
-            Object.defineProperty(Auth.prototype, "method", {
-                get: $util.oneOfGetter($oneOfFields = ["assumableIamRole", "kubernetesServiceAccount"]),
-                set: $util.oneOfSetter($oneOfFields)
-            });
-
-            /**
-             * Creates a new Auth instance using the specified properties.
-             * @function create
-             * @memberof flyteidl.admin.Auth
-             * @static
-             * @param {flyteidl.admin.IAuth=} [properties] Properties to set
-             * @returns {flyteidl.admin.Auth} Auth instance
-             */
-            Auth.create = function create(properties) {
-                return new Auth(properties);
-            };
-
-            /**
-             * Encodes the specified Auth message. Does not implicitly {@link flyteidl.admin.Auth.verify|verify} messages.
-             * @function encode
-             * @memberof flyteidl.admin.Auth
-             * @static
-             * @param {flyteidl.admin.IAuth} message Auth message or plain object to encode
-             * @param {$protobuf.Writer} [writer] Writer to encode to
-             * @returns {$protobuf.Writer} Writer
-             */
-            Auth.encode = function encode(message, writer) {
-                if (!writer)
-                    writer = $Writer.create();
-                if (message.assumableIamRole != null && message.hasOwnProperty("assumableIamRole"))
-                    writer.uint32(/* id 1, wireType 2 =*/10).string(message.assumableIamRole);
-                if (message.kubernetesServiceAccount != null && message.hasOwnProperty("kubernetesServiceAccount"))
-                    writer.uint32(/* id 2, wireType 2 =*/18).string(message.kubernetesServiceAccount);
-                return writer;
-            };
-
-            /**
-             * Decodes an Auth message from the specified reader or buffer.
-             * @function decode
-             * @memberof flyteidl.admin.Auth
-             * @static
-             * @param {$protobuf.Reader|Uint8Array} reader Reader or buffer to decode from
-             * @param {number} [length] Message length if known beforehand
-             * @returns {flyteidl.admin.Auth} Auth
-             * @throws {Error} If the payload is not a reader or valid buffer
-             * @throws {$protobuf.util.ProtocolError} If required fields are missing
-             */
-            Auth.decode = function decode(reader, length) {
-                if (!(reader instanceof $Reader))
-                    reader = $Reader.create(reader);
-                let end = length === undefined ? reader.len : reader.pos + length, message = new $root.flyteidl.admin.Auth();
-                while (reader.pos < end) {
-                    let tag = reader.uint32();
-                    switch (tag >>> 3) {
-                    case 1:
-                        message.assumableIamRole = reader.string();
-                        break;
-                    case 2:
-                        message.kubernetesServiceAccount = reader.string();
-                        break;
-                    default:
-                        reader.skipType(tag & 7);
-                        break;
-                    }
-                }
-                return message;
-            };
-
-            /**
-             * Verifies an Auth message.
-             * @function verify
-             * @memberof flyteidl.admin.Auth
-             * @static
-             * @param {Object.<string,*>} message Plain object to verify
-             * @returns {string|null} `null` if valid, otherwise the reason why it is not
-             */
-            Auth.verify = function verify(message) {
-                if (typeof message !== "object" || message === null)
-                    return "object expected";
-                let properties = {};
-                if (message.assumableIamRole != null && message.hasOwnProperty("assumableIamRole")) {
-                    properties.method = 1;
-                    if (!$util.isString(message.assumableIamRole))
-                        return "assumableIamRole: string expected";
-                }
-                if (message.kubernetesServiceAccount != null && message.hasOwnProperty("kubernetesServiceAccount")) {
-                    if (properties.method === 1)
-                        return "method: multiple values";
-                    properties.method = 1;
-                    if (!$util.isString(message.kubernetesServiceAccount))
-                        return "kubernetesServiceAccount: string expected";
-                }
-                return null;
-            };
-
-            return Auth;
         })();
 
         admin.LaunchPlanSpec = (function() {
