+++ gen/pb-protodoc/flyteidl/admin/execution.proto.rst
@@ -568,7 +568,8 @@ of an execution as it progresses across phase changes..
     "notifications": "{...}",
     "disable_all": "...",
     "labels": "{...}",
-    "annotations": "{...}"
+    "annotations": "{...}",
+    "auth": "{...}"
   }
 
 .. _api_field_flyteidl.admin.ExecutionSpec.launch_plan:
@@ -621,6 +622,12 @@ annotations
   (:ref:`flyteidl.admin.Annotations <api_msg_flyteidl.admin.Annotations>`) Annotations to apply to the execution resource.
   
   
+.. _api_field_flyteidl.admin.ExecutionSpec.auth:
+
+auth
+  (:ref:`flyteidl.admin.Auth <api_msg_flyteidl.admin.Auth>`) Optional: auth override to apply this execution.
+  
+  
 
 
 .. _api_msg_flyteidl.admin.ExecutionTerminateRequest:
@@ -628,7 +635,7 @@ annotations
 flyteidl.admin.ExecutionTerminateRequest
 ----------------------------------------
 
-`[flyteidl.admin.ExecutionTerminateRequest proto] <https://github.com/lyft/flyteidl/blob/master/protos/flyteidl/admin/execution.proto#L243>`_
+`[flyteidl.admin.ExecutionTerminateRequest proto] <https://github.com/lyft/flyteidl/blob/master/protos/flyteidl/admin/execution.proto#L246>`_
 
 Request to terminate an in-progress execution.  This action is irreversible.
 If an execution is already terminated, this request will simply be a no-op.
@@ -662,7 +669,7 @@ cause
 flyteidl.admin.ExecutionTerminateResponse
 -----------------------------------------
 
-`[flyteidl.admin.ExecutionTerminateResponse proto] <https://github.com/lyft/flyteidl/blob/master/protos/flyteidl/admin/execution.proto#L251>`_
+`[flyteidl.admin.ExecutionTerminateResponse proto] <https://github.com/lyft/flyteidl/blob/master/protos/flyteidl/admin/execution.proto#L254>`_
 
 
 .. code-block:: json
@@ -677,7 +684,7 @@ flyteidl.admin.ExecutionTerminateResponse
 flyteidl.admin.WorkflowExecutionGetDataRequest
 ----------------------------------------------
 
-`[flyteidl.admin.WorkflowExecutionGetDataRequest proto] <https://github.com/lyft/flyteidl/blob/master/protos/flyteidl/admin/execution.proto#L256>`_
+`[flyteidl.admin.WorkflowExecutionGetDataRequest proto] <https://github.com/lyft/flyteidl/blob/master/protos/flyteidl/admin/execution.proto#L259>`_
 
 Request structure to fetch inputs and output urls for an execution.
 
@@ -700,7 +707,7 @@ id
 flyteidl.admin.WorkflowExecutionGetDataResponse
 -----------------------------------------------
 
-`[flyteidl.admin.WorkflowExecutionGetDataResponse proto] <https://github.com/lyft/flyteidl/blob/master/protos/flyteidl/admin/execution.proto#L262>`_
+`[flyteidl.admin.WorkflowExecutionGetDataResponse proto] <https://github.com/lyft/flyteidl/blob/master/protos/flyteidl/admin/execution.proto#L265>`_
 
 Response structure for WorkflowExecutionGetDataRequest which contains inputs and outputs for an execution.
