+++ gen/pb-protodoc/flyteidl/admin/launch_plan.proto.rst
@@ -117,46 +117,12 @@ token
   
 
 
-.. _api_msg_flyteidl.admin.Auth:
-
-flyteidl.admin.Auth
--------------------
-
-`[flyteidl.admin.Auth proto] <https://github.com/lyft/flyteidl/blob/master/protos/flyteidl/admin/launch_plan.proto#L58>`_
-
-Defines permissions associated with executions created by this launch plan spec.
-
-.. code-block:: json
-
-  {
-    "assumable_iam_role": "...",
-    "kubernetes_service_account": "..."
-  }
-
-.. _api_field_flyteidl.admin.Auth.assumable_iam_role:
-
-assumable_iam_role
-  (`string <https://developers.google.com/protocol-buffers/docs/proto#scalar>`_) 
-  
-  
-  Only one of :ref:`assumable_iam_role <api_field_flyteidl.admin.Auth.assumable_iam_role>`, :ref:`kubernetes_service_account <api_field_flyteidl.admin.Auth.kubernetes_service_account>` may be set.
-  
-.. _api_field_flyteidl.admin.Auth.kubernetes_service_account:
-
-kubernetes_service_account
-  (`string <https://developers.google.com/protocol-buffers/docs/proto#scalar>`_) 
-  
-  
-  Only one of :ref:`assumable_iam_role <api_field_flyteidl.admin.Auth.assumable_iam_role>`, :ref:`kubernetes_service_account <api_field_flyteidl.admin.Auth.kubernetes_service_account>` may be set.
-  
-
-
 .. _api_msg_flyteidl.admin.LaunchPlanSpec:
 
 flyteidl.admin.LaunchPlanSpec
 -----------------------------
 
-`[flyteidl.admin.LaunchPlanSpec proto] <https://github.com/lyft/flyteidl/blob/master/protos/flyteidl/admin/launch_plan.proto#L67>`_
+`[flyteidl.admin.LaunchPlanSpec proto] <https://github.com/lyft/flyteidl/blob/master/protos/flyteidl/admin/launch_plan.proto#L58>`_
 
 User-provided launch plan definition and configuration values.
 
@@ -228,7 +194,7 @@ auth
 flyteidl.admin.LaunchPlanClosure
 --------------------------------
 
-`[flyteidl.admin.LaunchPlanClosure proto] <https://github.com/lyft/flyteidl/blob/master/protos/flyteidl/admin/launch_plan.proto#L96>`_
+`[flyteidl.admin.LaunchPlanClosure proto] <https://github.com/lyft/flyteidl/blob/master/protos/flyteidl/admin/launch_plan.proto#L87>`_
 
 Values computed by the flyte platform after launch plan registration.
 These include expected_inputs required to be present in a CreateExecutionRequest
@@ -281,7 +247,7 @@ updated_at
 flyteidl.admin.LaunchPlanMetadata
 ---------------------------------
 
-`[flyteidl.admin.LaunchPlanMetadata proto] <https://github.com/lyft/flyteidl/blob/master/protos/flyteidl/admin/launch_plan.proto#L115>`_
+`[flyteidl.admin.LaunchPlanMetadata proto] <https://github.com/lyft/flyteidl/blob/master/protos/flyteidl/admin/launch_plan.proto#L106>`_
 
 Additional launch plan attributes included in the LaunchPlanSpec not strictly required to launch
 the reference workflow.
@@ -312,7 +278,7 @@ notifications
 flyteidl.admin.LaunchPlanUpdateRequest
 --------------------------------------
 
-`[flyteidl.admin.LaunchPlanUpdateRequest proto] <https://github.com/lyft/flyteidl/blob/master/protos/flyteidl/admin/launch_plan.proto#L124>`_
+`[flyteidl.admin.LaunchPlanUpdateRequest proto] <https://github.com/lyft/flyteidl/blob/master/protos/flyteidl/admin/launch_plan.proto#L115>`_
 
 Request to set the referenced launch plan state to the configured value.
 
@@ -342,7 +308,7 @@ state
 flyteidl.admin.LaunchPlanUpdateResponse
 ---------------------------------------
 
-`[flyteidl.admin.LaunchPlanUpdateResponse proto] <https://github.com/lyft/flyteidl/blob/master/protos/flyteidl/admin/launch_plan.proto#L133>`_
+`[flyteidl.admin.LaunchPlanUpdateResponse proto] <https://github.com/lyft/flyteidl/blob/master/protos/flyteidl/admin/launch_plan.proto#L124>`_
 
 Purposefully empty, may be populated in the future.
 
@@ -358,7 +324,7 @@ Purposefully empty, may be populated in the future.
 flyteidl.admin.ActiveLaunchPlanRequest
 --------------------------------------
 
-`[flyteidl.admin.ActiveLaunchPlanRequest proto] <https://github.com/lyft/flyteidl/blob/master/protos/flyteidl/admin/launch_plan.proto#L137>`_
+`[flyteidl.admin.ActiveLaunchPlanRequest proto] <https://github.com/lyft/flyteidl/blob/master/protos/flyteidl/admin/launch_plan.proto#L128>`_
 
 Represents a request struct for finding an active launch plan for a given NamedEntityIdentifier
 
@@ -380,7 +346,7 @@ id
 flyteidl.admin.ActiveLaunchPlanListRequest
 ------------------------------------------
 
-`[flyteidl.admin.ActiveLaunchPlanListRequest proto] <https://github.com/lyft/flyteidl/blob/master/protos/flyteidl/admin/launch_plan.proto#L142>`_
+`[flyteidl.admin.ActiveLaunchPlanListRequest proto] <https://github.com/lyft/flyteidl/blob/master/protos/flyteidl/admin/launch_plan.proto#L133>`_
 
 Represents a request structure to list active launch plans within a project/domain.
