+++ gen/pb_python/flyteidl/service/flyteadmin/flyteadmin/models/admin_execution_spec.py
@@ -17,6 +17,7 @@ import re  # noqa: F401
 import six
 
 from flyteadmin.models.admin_annotations import AdminAnnotations  # noqa: F401,E501
+from flyteadmin.models.admin_auth import AdminAuth  # noqa: F401,E501
 from flyteadmin.models.admin_execution_metadata import AdminExecutionMetadata  # noqa: F401,E501
 from flyteadmin.models.admin_labels import AdminLabels  # noqa: F401,E501
 from flyteadmin.models.admin_notification_list import AdminNotificationList  # noqa: F401,E501
@@ -44,7 +45,8 @@ class AdminExecutionSpec(object):
         'notifications': 'AdminNotificationList',
         'disable_all': 'bool',
         'labels': 'AdminLabels',
-        'annotations': 'AdminAnnotations'
+        'annotations': 'AdminAnnotations',
+        'auth': 'AdminAuth'
     }
 
     attribute_map = {
@@ -54,10 +56,11 @@ class AdminExecutionSpec(object):
         'notifications': 'notifications',
         'disable_all': 'disable_all',
         'labels': 'labels',
-        'annotations': 'annotations'
+        'annotations': 'annotations',
+        'auth': 'auth'
     }
 
-    def __init__(self, launch_plan=None, inputs=None, metadata=None, notifications=None, disable_all=None, labels=None, annotations=None):  # noqa: E501
+    def __init__(self, launch_plan=None, inputs=None, metadata=None, notifications=None, disable_all=None, labels=None, annotations=None, auth=None):  # noqa: E501
         """AdminExecutionSpec - a model defined in Swagger"""  # noqa: E501
 
         self._launch_plan = None
@@ -67,6 +70,7 @@ class AdminExecutionSpec(object):
         self._disable_all = None
         self._labels = None
         self._annotations = None
+        self._auth = None
         self.discriminator = None
 
         if launch_plan is not None:
@@ -83,6 +87,8 @@ class AdminExecutionSpec(object):
             self.labels = labels
         if annotations is not None:
             self.annotations = annotations
+        if auth is not None:
+            self.auth = auth
 
     @property
     def launch_plan(self):
@@ -238,6 +244,29 @@ class AdminExecutionSpec(object):
         """
 
         self._annotations = annotations
+
+    @property
+    def auth(self):
+        """Gets the auth of this AdminExecutionSpec.  # noqa: E501
+
+        Optional: auth override to apply this execution.  # noqa: E501
+
+        :return: The auth of this AdminExecutionSpec.  # noqa: E501
+        :rtype: AdminAuth
+        """
+        return self._auth
+
+    @auth.setter
+    def auth(self, auth):
+        """Sets the auth of this AdminExecutionSpec.
+
+        Optional: auth override to apply this execution.  # noqa: E501
+
+        :param auth: The auth of this AdminExecutionSpec.  # noqa: E501
+        :type: AdminAuth
+        """
+
+        self._auth = auth
 
     def to_dict(self):
         """Returns the model properties as a dict"""
