+++ protos/flyteidl/admin/execution.proto
@@ -234,6 +234,9 @@ message ExecutionSpec {
     // Annotations to apply to the execution resource.
     Annotations annotations   = 8;
 
+    // Optional: auth override to apply this execution.
+    Auth auth                 = 16;
+
 }
 
 // Request to terminate an in-progress execution.  This action is irreversible.
