+++ protos/flyteidl/admin/launch_plan.proto
@@ -55,15 +55,6 @@ message LaunchPlanList {
     string token                     = 2;
 }
 
-// Defines permissions associated with executions created by this launch plan spec.
-message Auth {
-  oneof method {
-    string assumable_iam_role         = 1;
-    string kubernetes_service_account = 2;
-  }
-}
-
-
 // User-provided launch plan definition and configuration values.
 message LaunchPlanSpec {
     // Reference to the Workflow template that the launch plan references
